compliance-continuum-server/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
supabase>=2.15.0,<3.0.0
redis==5.0.4
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.1

httpx>=0.24,<0.27
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
requests==2.31.0
python-dotenv